Bootstrap shutdown (#112)
Allow user to know when `ClientBootstrap` is done shutting down its behind-the-scenes resources.

User may register a `std::function<void()>` that fires upon shutdown complete. I considered offering a `std::future` to users, but realized you can only do blocking wait()/get() operations on a c++ future, you can't register a callback to fire (a feature available in Python and Java's future primitives). So if a user wants to use `std::future` they can hook that up themselves with the callback (I did this in a test).

Also added an `EnableBlockingShutdown()` function. I was looking around at how async-shutdown was handled in other `aws-crt-cpp` classes and saw this feature on the `HttpClientConnectionManager`. It's simplifies things when the IO classes are created on the main thread's stack (all of our tests and samples).

I chose to make these functions you call on the constructed class, rather than constructor arguments, or fields in an "Options" struct for 2 reasons: 1) This class wasn't using an Options struct before, so it felt weird to add one now. 2) If these were overloaded constructor args, I feared a user would enable the "blocking shutdown" by copy/pasting a constructor with `..., true, ...` in it and later have no idea why they occasionally dead-locked.