Only using extra warning and werror flags on Debug build type (#246)
Currently we force Werror and extra warnings in every type of build.
This is a pain point to package mantainers and cannot be deactivated
easily. Then, when a compiler or package manager we didn't explicitly
tested against, for example when new compiler version is launched.
Release build may fail to users. We are going to test our build in
Debug type of build to catch this warnings in CI, but they not going
to fail release build for users.