Version fix (#412)
**Issue:**

`AWS_CRT_CPP_VERSION` mistakenly reports the git tag of the **PARENT** project when built via CMake's add_subdirectory()

**Complication:**
This bug was accidentally helping the IoT SDK report its own version in the the [MqttClientConnectionConfigBuilder metrics string](https://github.com/awslabs/aws-crt-cpp/blob/main/source/iot/MqttClient.cpp#L476-L490). This code lives in the CRT, but probably should live in the IoT SDK, but we want it to report the IoT SDK's version, and apparently it has been doing that all this time, accidentally! and we don't want to break that.

**Description of changes:**
1) Rely on a simple `VERSION` file in the repo to determine version at build time.
2) Fix git calls so they report THIS project's hash, not the parent project's
3) New build option `AWS_IOT_SDK_VERSION` lets the IoT SDK override the version reported by the MqttClientConnectionConfigBuilder metrics string